# backend/swagger/swagger.json
  {
      "openapi": "3.0.0",
        "info": {
          "title": "Basic GameBuster API",
          "version": "1.0.0",
          "description": "A basic API"
        },
      "servers": [
        {
          "url": "http://localhost:4000/api",
          "description": "Localhost Server for testing the GameBuster API"
        }
      ],
      "paths": {
...
      "/one/{_id}": {
        "get":{
          "tags":[
            "peliculas"
          ],
          "summary": "Get a specify Movie",
          "description": "Busca una pelicula en especifico",
          "parameters":[
            {
              "in": "path",
              "name": "_id",
              "required": true,
              "schema": {
                "type": "string",
                "description": "Id de la pelicua a buscar"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "items": {
                      "$ref": "no schema defined"
                    },
                    "example": {
                      "peliculas": {
                        "pelicula": "FNAF",
                        "stock": 30,
                        "precio": 7.99,
                        "categoria": "Fantasía/Aventura"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "NOT OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "no schema defined"
                    },
                    "example": {
                      "error": "Bad Request (could change)"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    }